starts to build out Button usage architecture
